+++ cmp/grid/columns/Column.js
@@ -113,12 +113,12 @@ export class Column {
      * @param {(ExportFormat|function)} [c.exportFormat] - structured format string for Excel-based
      *      exports, or a function to produce one. {@see ExportFormat}
      * @param {number} [c.exportWidth] - width in characters for Excel-based exports. Typically
-     *     used
-     *      with ExportFormat.LONG_TEXT to enable text wrapping.
+     *     used with ExportFormat.LONG_TEXT to enable text wrapping.
      * @param {(boolean|Column~tooltipFn)} [c.tooltip] - 'true' displays the raw value, or
-     *      tool tip function, which is based on AG Grid tooltip callback.
+     *      tooltip function, which is based on AG Grid tooltip callback. Incompatible with
+     *      `tooltipElement`.
      * @param {Column~tooltipElementFn} [c.tooltipElement] - function which returns a React
-     *     component.
+     *     component to display as a tooltip. Will take precedence over `tooltip`.
      * @param {boolean} [c.excludeFromExport] - true to drop this column from a file export.
      * @param {boolean} [c.autosizable] - allow autosizing this column.
      * @param {boolean} [c.autosizeIncludeHeader] - true to include the header width when
@@ -293,6 +293,10 @@ export class Column {
 
         this.tooltip = tooltip;
         this.tooltipElement = tooltipElement;
+        warnIf(
+            tooltip && tooltipElement,
+            `Column specified with both tooltip && tooltipElement. Tooltip will be ignored. [colId=${this.colId}]`
+        );
 
         this.editable = editable;
         this.setValueFn = withDefault(setValueFn, this.defaultSetValueFn);
@@ -396,13 +400,11 @@ export class Column {
         }
 
         // Tooltip Handling
-        const {tooltip, tooltipElement} = this;
-        if (tooltip || tooltipElement) {
-            ret.tooltipValueGetter = (agParams) => {
-                const value = isFunction(tooltip) ?
-                    tooltip(agParams.value, {record: agParams.data, column: me, gridModel, agParams}) :
-                    agParams.value;
+        const {tooltip, tooltipElement} = this,
+            tooltipSpec = tooltipElement ?? tooltip;
 
+        if (tooltipSpec) {
+            ret.tooltipValueGetter = ({value}) => {
                 // Note that due to a known AgGrid issue, a tooltip must always return a value
                 // or risk not showing when the value later changes.
                 // See https://github.com/xh/hoist-react/issues/2058
@@ -419,14 +421,16 @@ export class Column {
 
                     if (location === 'header') return div(me.headerTooltip);
 
+                    // ag-Grid cmp gets escaped value, lookup raw value from record instead
                     const record = api.getDisplayedRowAtIndex(rowIndex)?.data;
-                    if (record && tooltipElement) {
-                        // ag-Grid cmp gets escaped value, lookup raw value from record instead
-                        const {store} = record,
-                            value = me.getValueFn({record, column: me, gridModel, agParams, store});
-                        return tooltipElement(value, {record, column: me, gridModel, agParams});
-                    }
-                    return agParams.value === '*EMPTY*' ? null : agParams.value;
+                    if (!record) return null;
+
+                    const {store} = record,
+                        value = me.getValueFn({record, column: me, gridModel, agParams, store});
+
+                    return isFunction(tooltipSpec) ?
+                        tooltipSpec(value, {record, column: me, gridModel, agParams}) :
+                        value;
                 }
             };
         }
@@ -651,7 +655,7 @@ export function getAgHeaderClassFn(column) {
  * @callback Column~tooltipElementFn - function for a grid cell tooltip returning a React element.
  * @param {*} value - tooltip value.
  * @param {TooltipMetadata} metadata - additional data about the column and row.
- * @return {Element} - the React element to show as a toooltip.
+ * @return {Element} - the React element to show as a tooltip.
  */
 
 /**